--- a/benchmarks/basic.bench.js
+++ b/benchmarks/basic.bench.js
@@ -10,7 +10,7 @@ var dest = fs.createWriteStream('/dev/null')
 var loglevel = require('./loglevelMock')(dest)
 var plog = pino(dest)
 delete require.cache[require.resolve('../')]
-var plogExtreme = require('../')({extreme: true}, dest)
+var plogExtreme = require('../')(pino.extreme('/dev/null'))
 
 process.env.DEBUG = 'dlog'
 var debug = require('debug')