--- a/benchmarks/child.bench.js
+++ b/benchmarks/child.bench.js
@@ -8,7 +8,7 @@ var fs = require('fs')
 var dest = fs.createWriteStream('/dev/null')
 var plog = pino(dest).child({ a: 'property' })
 delete require.cache[require.resolve('../')]
-var plogExtreme = require('../')({extreme: true}, dest).child({ a: 'property' })
+var plogExtreme = require('../')(pino.extreme('/dev/null')).child({ a: 'property' })
 
 var max = 10
 var blog = bunyan.createLogger({