--- a/benchmarks/childCreation.bench.js
+++ b/benchmarks/childCreation.bench.js
@@ -8,7 +8,7 @@ var fs = require('fs')
 var dest = fs.createWriteStream('/dev/null')
 var plog = pino(dest)
 delete require.cache[require.resolve('../')]
-var plogExtreme = require('../')({extreme: true}, dest)
+var plogExtreme = require('../')(pino.extreme('/dev/null'))
 
 var max = 10
 var blog = bunyan.createLogger({