--- a/benchmarks/deep-object.bench.js
+++ b/benchmarks/deep-object.bench.js
@@ -9,11 +9,11 @@ var fs = require('fs')
 var dest = fs.createWriteStream('/dev/null')
 var plog = pino(dest)
 delete require.cache[require.resolve('../')]
-var plogExtreme = require('../')({extreme: true}, dest)
+var plogExtreme = require('../')(pino.extreme('/dev/null'))
 delete require.cache[require.resolve('../')]
 var plogUnsafe = require('../')({safe: false}, dest)
 delete require.cache[require.resolve('../')]
-var plogUnsafeExtreme = require('../')({extreme: true, safe: false}, dest)
+var plogUnsafeExtreme = require('../')({safe: false}, pino.extreme('/dev/null'))
 
 var loglevel = require('./loglevelMock')(dest)