--- a/benchmarks/just-pino.bench.js
+++ b/benchmarks/just-pino.bench.js
@@ -6,7 +6,7 @@ var fs = require('fs')
 var dest = fs.createWriteStream('/dev/null')
 var plog = pino(dest)
 delete require.cache[require.resolve('../')]
-var plogExtreme = require('../')({extreme: true}, dest)
+var plogExtreme = require('../')(pino.extreme('/dev/null'))
 var max = 10
 
 var run = bench([