--- a/lib/tools.js
+++ b/lib/tools.js
@@ -74,12 +74,6 @@ function defineLevelsProperty (onObject) {
   Object.defineProperty(onObject.levels.labels, Infinity, {value: 'silent'})
 }
 
-function streamIsBlockable (s) {
-  if (s.hasOwnProperty('_handle') && s._handle.hasOwnProperty('fd') && s._handle.fd) return true
-  if (s.hasOwnProperty('fd') && s.fd) return true
-  return false
-}
-
 function countInterp (s, i) {
   var n = 0
   var pos = 0
@@ -135,6 +129,5 @@ module.exports = {
   applyOptions: applyOptions,
   asMetaWrapper: asMetaWrapper,
   defineLevelsProperty: defineLevelsProperty,
-  streamIsBlockable: streamIsBlockable,
   genLog: genLog
 }