--- a/test/fixtures/events/destination.js
+++ b/test/fixtures/events/destination.js
@@ -17,6 +17,6 @@ var log = pino(pino.destination())
 log.info('h')
 
 function foo () {
-  setTimeout(foo, 50)
+  setTimeout(foo, 500)
 }
 foo()