--- a/test/fixtures/events/no-destination.js
+++ b/test/fixtures/events/no-destination.js
@@ -1,0 +1,21 @@
+'use strict'
+
+// we need stable data to run our tests
+global.process = { __proto__: process, pid: 123456 }
+Date.now = function () { return 1459875739796 }
+require('os').hostname = function () { return 'abcdefghijklmnopqr' }
+
+if (process.listenerCount('SIGHUP') > 0) {
+  // needed because of a hook added by code coverage
+  process.removeAllListeners('SIGHUP')
+}
+
+var pino = require(require.resolve('./../../../'))
+
+var log = pino()
+log.info('h')
+
+function foo () {
+  setTimeout(foo, 50)
+}
+foo()