--- a/test/fixtures/events/no-destination.js
+++ b/test/fixtures/events/no-destination.js
@@ -16,6 +16,6 @@ var log = pino()
 log.info('h')
 
 function foo () {
-  setTimeout(foo, 50)
+  setTimeout(foo, 500)
 }
 foo()