--- a/test/fixtures/events/sighup-no-handler.js
+++ b/test/fixtures/events/sighup-no-handler.js
@@ -13,7 +13,7 @@ if (process.listenerCount('SIGHUP') > 0) {
 var pino = require(require.resolve('./../../../'))
 
 // extreme mode
-var log = pino({extreme: true})
+var log = pino(pino.extreme())
 log.info('h')
 
 function foo () {