--- a/test/fixtures/events/sighup-with-handler.js
+++ b/test/fixtures/events/sighup-with-handler.js
@@ -2,7 +2,7 @@ global.process = { __proto__: process, pid: 123456 }
 Date.now = function () { return 1459875739796 }
 require('os').hostname = function () { return 'abcdefghijklmnopqr' }
 var pino = require(require.resolve('./../../../'))
-var log = pino({extreme: true})
+var log = pino(pino.extreme())
 log.info('h')
 
 process.on('SIGHUP', function () {