--- a/test/fixtures/extreme.js
+++ b/test/fixtures/extreme.js
@@ -2,5 +2,5 @@ global.process = { __proto__: process, pid: 123456 }
 Date.now = function () { return 1459875739796 }
 require('os').hostname = function () { return 'abcdefghijklmnopqr' }
 var pino = require(require.resolve('./../../'))
-var extreme = pino({extreme: true})
+var extreme = pino(pino.extreme())
 extreme.info('h')