--- a/test/fixtures/extreme_child.js
+++ b/test/fixtures/extreme_child.js
@@ -2,5 +2,5 @@ global.process = { __proto__: process, pid: 123456 }
 Date.now = function () { return 1459875739796 }
 require('os').hostname = function () { return 'abcdefghijklmnopqr' }
 var pino = require(require.resolve('./../../'))
-var extreme = pino({extreme: true}).child({ hello: 'world' })
+var extreme = pino(pino.extreme()).child({ hello: 'world' })
 extreme.info('h')