This still needs some work but I've opened it get some feedbacks.

Fixes https://github.com/pinojs/pino/issues/341.
Fixes #393.