@mcollina 2018-04-13
This has a perf regression on extreme mode, I will investigate.

master:

```
benchBunyan*10000: 834.808ms
benchWinston*10000: 623.330ms
benchBole*10000: 336.978ms
benchDebug*10000: 368.957ms
benchLogLevel*10000: 464.211ms
benchPino*10000: 331.520ms
benchPinoExtreme*10000: 238.182ms
benchBunyan*10000: 715.397ms
benchWinston*10000: 538.094ms
benchBole*10000: 313.578ms
benchDebug*10000: 327.992ms
benchLogLevel*10000: 396.778ms
benchPino*10000: 306.623ms
benchPinoExtreme*10000: 217.418ms
```

This PR:

```
$ node benchmarks/basic.bench.js
benchBunyan*10000: 906.333ms
benchWinston*10000: 626.493ms
benchBole*10000: 330.269ms
benchDebug*10000: 352.773ms
benchLogLevel*10000: 422.380ms
benchPino*10000: 319.075ms
benchPinoExtreme*10000: 173.527ms
benchBunyan*10000: 734.764ms
benchWinston*10000: 536.962ms
benchBole*10000: 309.273ms
benchDebug*10000: 339.026ms
benchLogLevel*10000: 409.393ms
benchPino*10000: 318.122ms
benchPinoExtreme*10000: 150.465ms
```