@davidmarkclements 2018-04-13
doc updates will be separate PR? 

if so I don't think we should merge without updating docs (in case it gets forgotten)