@mcollina 2018-04-14
This PR:

```
$ node benchmarks/basic.bench.js
benchBunyan*10000: 849.859ms
benchWinston*10000: 631.939ms
benchBole*10000: 309.742ms
benchDebug*10000: 347.677ms
benchLogLevel*10000: 459.549ms
benchPino*10000: 304.755ms
benchPinoDestination*10000: 219.132ms
benchPinoExtreme*10000: 172.622ms
benchBunyan*10000: 727.658ms
benchWinston*10000: 501.120ms
benchBole*10000: 316.504ms
benchDebug*10000: 332.890ms
benchLogLevel*10000: 406.898ms
benchPino*10000: 295.664ms
benchPinoDestination*10000: 223.564ms
benchPinoExtreme*10000: 160.127ms
```

Master

```
$ node benchmarks/basic.bench.js
benchBunyan*10000: 843.972ms
benchWinston*10000: 610.107ms
benchBole*10000: 324.121ms
benchDebug*10000: 352.105ms
benchLogLevel*10000: 442.383ms
benchPino*10000: 331.777ms
benchPinoExtreme*10000: 167.160ms
benchBunyan*10000: 716.411ms
benchWinston*10000: 604.688ms
benchBole*10000: 376.154ms
benchDebug*10000: 381.350ms
benchLogLevel*10000: 426.684ms
benchPino*10000: 301.478ms
benchPinoExtreme*10000: 150.016ms
```

Note that there is the new `destination` object based on SonicBoom that gives us a nice boost.