@mcollina 2018-04-16
@jsumners the events tests are not passing on Node 6 on Linux. However they are on Mac. Can you have a look?

```
 # Subtest: terminates on SIGHUP when no other handlers registered
        1..2
        not ok 1 - should be equal
          ---
          found: null
          wanted: 0
          compare: ===
          at:
            line: 80
            column: 7
            file: test/events.test.js
            type: ChildProcess
          stack: |
            ChildProcess.<anonymous> (test/events.test.js:80:7)
          source: |
            t.is(code, 0)
          ...
        
        not ok 2 - should not be equal
          ---
          found: null
          doNotWant: null
          compare: '!=='
          at:
            line: 84
            column: 7
            file: test/events.test.js
            type: ChildProcess
          stack: |
            ChildProcess.<anonymous> (test/events.test.js:84:7)
          source: |
            t.notEqual(output.match(/"msg":"h"/), null)
          ...
        
        # failed 2 of 2 tests
    not ok 4 - terminates on SIGHUP when no other handlers registered # time=1167.903ms
```