@jsumners 2018-04-16
I'll have to setup a Linux environment to test. That _could_ happen today, but we'll see.